Created basic view by type, above minimum distance, and above minimum duration.

+++ src/Activity.java
@@ -30,7 +30,7 @@ public class Activity implements Comparable<Activity> {
         this.type = type;
     }
 
-    public double getDuration() {
+    public int getDuration() {
         return duration;
     }
